+++ bookstoreaggregator/books/fixtures/initial_data.json
