+++ bookstoreaggregator/books/management/commands/list_books.py
