+++ bookstoreaggregator/books/templates/books/book_list.html
